+++ src/Login.tsx
@@ -1,4 +1,4 @@
-import { Box, Button, Card, Container, TextField, Typography } from "@mui/material";
+import { Alert, Box, Button, Card, Container, TextField, Typography } from "@mui/material";
 import { useEffect, useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import Appbar from "./Appbar";
@@ -7,6 +7,7 @@ function Login({ setUser }: { setUser: any }) {
    const navigate = useNavigate();
    const [username, setUsername] = useState<string>("");
    const [password, setPassword] = useState<string>("");
+   const [alertMessage, setAlertMessage] = useState<string>("");
 
    useEffect(() => {
       setUser(false);
@@ -27,12 +28,23 @@ function Login({ setUser }: { setUser: any }) {
       });
 
       const data = await response.json();
-      if (data.status === "error") console.log("there was an error");
+      if (data.status === "error") setAlertMessage(data.message);
       else if (data.status === "ok") {
-         console.log(data.user);
          setUser(data.user);
          localStorage.setItem("user", JSON.stringify(data.user));
          navigate("/jomaker/app");
+      }
+   }
+
+   function alert() {
+      if (alertMessage) {
+         return (
+            <Alert severity="error" sx={{ mt: "10px", width: "90%" }}>
+               {alertMessage}
+            </Alert>
+         );
+      } else {
+         return <></>;
       }
    }
 
@@ -91,6 +103,7 @@ function Login({ setUser }: { setUser: any }) {
                <Link to="/jomaker/offline" style={{ width: "100%", display: "flex", justifyContent: "flex-end" }}>
                   <Typography variant="subtitle2">Use Offline Version</Typography>
                </Link>
+               {alert()}
             </Card>
          </Container>
       </Box>